ver.0.3.4
Optimised some of my scripts, implemented new buzzing sound for lights.

+++ Assets/Scripts/Menu/MainMenuManager.cs
@@ -6,11 +6,7 @@ using UnityEngine.Events;
 using UnityEngine.UI;
 using System;
 using System.IO;
-public enum SliderValue
-{
-    master, soundFX, music, ui, res
-}
-
+public enum SliderValue {master, soundFX, music, ui, res}
 
 [System.Serializable]
 public class MenuButton
@@ -101,20 +97,15 @@ public class MainMenuManager : MonoBehaviour
 
     private void Start()
     {
-        //Cursor.lockState = CursorLockMode.Locked;
-        //Cursor.visible = false;
+        Cursor.lockState = CursorLockMode.Locked;
+        Cursor.visible = false;
         
         SettingsManager.LoadResolutions(false);
 
         UpdateMenuFonts();
         UpdateHighlightDisplay();
 
-        float fillAmount = 0f;
-        if (SettingsManager.firstBoot)
-        {
-            //MusicManager.ChangeTrack(splashMusic, true);
-            fillAmount = 1f;
-        }
+        float fillAmount = SettingsManager.firstBoot ? 1f : 0f;
 
         splashImage.fillAmount = fillAmount;
         StartCoroutine(ImageFill(splashImage, fillAmount));
@@ -143,7 +134,7 @@ public class MainMenuManager : MonoBehaviour
         {
             StartCoroutine(ImageFill(splashImage, 1f));
         }
-
+		
         if (vertInput > .5f)
         {
             UpdateSelection(-1);
@@ -158,7 +149,7 @@ public class MainMenuManager : MonoBehaviour
         {
             vertHeld = false;
         }
-
+		
         if (hozInput > .5f)
         {
             SliderUse(1);
@@ -174,21 +165,14 @@ public class MainMenuManager : MonoBehaviour
             hozHeld = false;
         }
 
-        if (Input.GetButtonDown("Submit"))
-        {
-            ActivateSelection();
-        }
+        if (Input.GetButtonDown("Submit"))	ActivateSelection();
 
-        if (Input.GetButtonDown("Cancel") && currentPageIndex != 0)
-        {
-            ChangePage(0);
-        }
+        if (Input.GetButtonDown("Cancel") && currentPageIndex != 0)	 ChangePage(0);
     }
 
     IEnumerator ImageFill(Image img, float fillAmount)
     {
         splashChanging = true;
-
         idleTime = 0f;
 
         while (img.fillAmount != fillAmount)
@@ -198,15 +182,9 @@ public class MainMenuManager : MonoBehaviour
         }
 
         displayingSplash = fillAmount == 1f;
-
-        if (displayingSplash)
-        {
-            MusicManager.ChangeTrack(splashMusic, true);
-        }
-        else
-        {
-            MusicManager.ChangeTrack(menuMusic, true);
-        }
+		
+		AudioClip track = displayingSplash ? splashMusic : menuMusic;
+		MusicManager.ChangeTrack(track, true);
 
         splashParent.SetActive(displayingSplash);
         splashChanging = false;
@@ -214,15 +192,8 @@ public class MainMenuManager : MonoBehaviour
 
     private void MenuHighlightMove(RectTransform targetRect)
     {
-        if(targetRect == null || !SettingsManager.data.menuHighlight)
-        {
-            return;
-        }
-
-        if(highlightImage == null)
-        {
-            CreateHighlightSprite();
-        }
+        if(targetRect == null || !SettingsManager.data.menuHighlight)	return;
+        if(highlightImage == null)	CreateHighlightSprite();
 
         highlightImage.transform.SetParent(targetRect);
         highlightImage.transform.localRotation = Quaternion.Euler(Vector3.zero);
@@ -240,15 +211,12 @@ public class MainMenuManager : MonoBehaviour
 
     void RemoveButtons()
     {
-        if(highlightImage != null && highlightImage.transform.parent != null)
+        if(highlightImage?.transform.parent != null)
         {
-            highlightImage.transform.SetParent(null);
+            highlightImage?.transform.SetParent(null);
         }
 
-        for (int i = 0; i < spawnedMenuButtons.Count; i++)
-        {
-            Destroy(spawnedMenuButtons[i]);
-        }
+        for (int i = 0; i < spawnedMenuButtons.Count; i++)	Destroy(spawnedMenuButtons[i]);
 
         spawnedMenuButtons.Clear();
     }
@@ -261,10 +229,7 @@ public class MainMenuManager : MonoBehaviour
 
         MenuPage curPage = Pages[currentPageIndex];
 
-        if(curPage.menuButtons.Count <= maxButtonsPerPage)
-        {
-            currentSubPageIndex = 0;
-        }
+        if(curPage.menuButtons.Count <= maxButtonsPerPage)	currentSubPageIndex = 0;
 
         int startIndex = currentSubPageIndex * maxButtonsPerPage;
         int endIndex = startIndex + maxButtonsPerPage;
@@ -327,11 +292,7 @@ public class MainMenuManager : MonoBehaviour
             {
                 if (SettingsManager.data.menuHighlight)
                 {
-                    if(highlightImage == null)
-                    {
-                        CreateHighlightSprite();
-                    }
-
+                    if(highlightImage == null)	CreateHighlightSprite();
                     MenuHighlightMove(tempButton.GetComponent<RectTransform>());
                 }
 
@@ -413,10 +374,7 @@ public class MainMenuManager : MonoBehaviour
 
     private void UpdateSlider(float fillValue, float max = 100f)
     {
-        if (sliderImage == null)
-        {
-            return;
-        }
+        if (sliderImage == null)	return;
 
         sliderImage.fillAmount = Mathf.Clamp(fillValue / max, 0f, 1f);
         sliderImage.color = Color.Lerp(Color.red, Color.green, sliderImage.fillAmount / 1f);
@@ -424,10 +382,7 @@ public class MainMenuManager : MonoBehaviour
 
     void SliderUse(int index)
     {
-        if (hozHeld || pageTurning)
-        {
-            return;
-        }
+        if (hozHeld || pageTurning)	return;
 
         MenuPage curPage = Pages[currentPageIndex];
         MenuButton curButton = curPage.menuButtons[curPage.currentButtonIndex];
@@ -500,10 +455,7 @@ public class MainMenuManager : MonoBehaviour
     {
         string path = Application.persistentDataPath + "/gamesettings.json";
 
-        if (File.Exists(path))
-        {
-            File.Delete(path);
-        }
+        if (File.Exists(path))	File.Delete(path);
 
         SettingsManager.data = new();
         SettingsManager.LoadSettings();
@@ -521,39 +473,26 @@ public class MainMenuManager : MonoBehaviour
 
     private void UpdateMenuFonts()
     {
-        menuButtonPrefab.GetComponent<TMP_Text>().font = defaultButtonFont;
-        pageTitleText.font = defaultButtonFont;
-        buttonDescText.font = defaultButtonFont;
-
-        if (SettingsManager.data.simpleFont)
-        {
-            menuButtonPrefab.GetComponent<TMP_Text>().font = simpleButtonFont;
-            pageTitleText.font = simpleButtonFont;
-            buttonDescText.font = simpleButtonFont;
-        }
+		menuButtonPrefab.GetComponent<TMP_Text>().font = SettingsManager.data.simpleFont ? simpleButtonFont : defaultButtonFont;
+		pageTitleText.font = SettingsManager.data.simpleFont ? simpleButtonFont : defaultButtonFont;
+		buttonDescText.font = SettingsManager.data.simpleFont ? simpleButtonFont : defaultButtonFont;
     }
 
     public void ToggleHighlight()
     {
         SettingsManager.data.menuHighlight = !SettingsManager.data.menuHighlight;
-
         UpdateHighlightDisplay();
     }
 
     private void UpdateHighlightDisplay()
     {
-        if (highlightImage != null)
-        {
-            highlightImage.enabled = SettingsManager.data.menuHighlight;
-        }
+		if(highlightImage == null) return;
+		highlightImage.enabled = SettingsManager.data.menuHighlight;
     }
 
     void UpdateSelection(int index)
     {
-        if (vertHeld)
-        {
-            return;
-        }
+        if (vertHeld)	return;
 
         MenuPage curPage = Pages[currentPageIndex];
 
@@ -566,7 +505,7 @@ public class MainMenuManager : MonoBehaviour
 
         if (curPage.currentButtonIndex > endIndex - 1)
         {
-            if(curPage.currentButtonIndex < curPage.menuButtons.Count)
+			if(curPage.currentButtonIndex < curPage.menuButtons.Count)
             {
                 currentSubPageIndex++;
             }
@@ -577,7 +516,7 @@ public class MainMenuManager : MonoBehaviour
         }
         else if (curPage.currentButtonIndex < startIndex)
         {
-            if(curPage.currentButtonIndex > 0)
+			if(curPage.currentButtonIndex > 0)
             {
                 currentSubPageIndex--;
             }
@@ -632,10 +571,7 @@ public class MainMenuManager : MonoBehaviour
     {
         pageImage.enabled = img != null && Pages[currentPageIndex].displayImage;
 
-        if (img == null)
-        {
-            return;
-        }
+        if (img == null)	return;
 
         pageImage.sprite = img;
         pageImage.color = colour;
@@ -657,20 +593,11 @@ public class MainMenuManager : MonoBehaviour
             JournalEntry currentEntry = journalObj.EntryContainer.Entries[i];
             MenuButton newEntry = new();
 
-            string buttonName = new string('?', currentEntry.Title.Length);
-            string buttonDesc = new string('?', currentEntry.Description.Length);
+			string buttonName = currentEntry.Owned ? currentEntry.Title : new string('?', currentEntry.Title.Length);
+			string buttonDesc = currentEntry.Owned ? currentEntry.Description : new string('?', currentEntry.Description.Length);
 
-            Sprite tempImage = currentEntry.missingImage;
-            Color tempColour = Color.black;
-
-            if (currentEntry.Owned)
-            {
-                buttonName = currentEntry.Title;
-                buttonDesc = currentEntry.Description;
-
-                tempImage = currentEntry.Picture;
-                tempColour = Color.white;
-            }
+			Sprite tempImage = currentEntry.Owned ? currentEntry.Picture : currentEntry.missingImage;
+            Color tempColour = currentEntry.Owned ? Color.white : Color.black;
 
             newEntry.buttonName = buttonName;
             newEntry.buttonDesc = buttonDesc;
@@ -691,27 +618,14 @@ public class MainMenuManager : MonoBehaviour
 
     public void ChangePage(int page)
     {
-        if (page > Pages.Count || page < 0)
-        {
-            return;
-        }
+        if (page > Pages.Count || page < 0)	return;
 
         currentSubPageIndex = 0;
 
-        if (leftPageDupe != null)
-        {
-            Destroy(leftPageDupe.gameObject);
-        }
-        if (rightPageDupe != null)
-        {
-            Destroy(rightPageDupe.gameObject);
-        }
+        if (leftPageDupe != null)	Destroy(leftPageDupe.gameObject);
+        if (rightPageDupe != null)	Destroy(rightPageDupe.gameObject);
 
-
-        if (highlightImage != null && highlightImage.transform.parent != null)
-        {
-            highlightImage.transform.SetParent(null);
-        }
+        if (highlightImage?.transform.parent != null)	highlightImage.transform.SetParent(null);
 
         StopAllCoroutines();
         StartCoroutine(PageAnimation(page));
@@ -797,10 +711,7 @@ public class MainMenuManager : MonoBehaviour
 
     void ActivateSelection()
     {
-        if (pageTurning)
-        {
-            return;
-        }
+        if (pageTurning)	return;
 
         MenuPage curMenu = Pages[currentPageIndex];
         MenuButton button = curMenu.menuButtons[curMenu.currentButtonIndex];
@@ -811,10 +722,7 @@ public class MainMenuManager : MonoBehaviour
             button.SelectEvent.Invoke();
             tempButtonSound = activateSound;
         }
-        if (button.playSounds)
-        {
-            AudioManager.PlayAudio(AudioType.ui, tempButtonSound, null, Vector2.zero, null, 1, 1, 0);
-        }
+        if (button.playSounds)	AudioManager.PlayAudio(AudioType.ui, tempButtonSound, null, Vector2.zero, null, 1, 1, 0);
     }
 
     public void LoadScene(string scene)
@@ -833,15 +741,9 @@ public class MainMenuManager : MonoBehaviour
     {
         string path = Application.persistentDataPath;
 
-        if(File.Exists(path + "/journal.json"))
-        {
-            File.Delete(path + "/journal.json");
-        }
+        if(File.Exists(path + "/journal.json"))	File.Delete(path + "/journal.json");
 
-        if (File.Exists(path + "/scores.json"))
-        {
-            File.Delete(path + "/scores.json");
-        }
+        if (File.Exists(path + "/scores.json"))	File.Delete(path + "/scores.json");
 
         ChangePage(0);
         LeaderboardManager.LoadScores();
